Draft of Number 2 and 3
Drafted Number 2.

Did something interesting for number 3:
We need to make "final predictions," not just have a bunch of models.
So the idea is ensemble learning:

Train a bunch of different models, and then predict an article is
"unpopular" if ANY of the models predict "unpopular".  While this won't
have the highest accuracy, it will reduce the number of "unpopular"
articles which we (wrongly) predict as "popular": This is the expensive
mistake that we want to avoid.